+++ src/actions/internal/certificate/generateCertificate.js
@@ -1,11 +1,15 @@
 import { getCertPath } from '../tools/certPaths.js';
 import { getMandatoryEnvValue } from '../../../utils/env.js';
-import { execute } from '../../../utils/execute.js';
+import { checkCommand, execute } from '../../../utils/execute.js';
 import { getLog } from '../../../utils/log.js';
+import { EnvironmentError } from '../../../utils/errors/index.js';
 
 const log = getLog('generateCertificate');
 
 export async function generateCertificate({ pwd }) {
+  if (!(await checkCommand({ pwd, commandBin: 'openssl' }))) {
+    throw new EnvironmentError(1659356205, 'You have to have openssl install to create cert');
+  }
   const { certPath, certKeyPath } = getCertPath({ pwd });
   const host = getMandatoryEnvValue('HOST');
   const proxyName = getMandatoryEnvValue('PROXY_NAME');