+++ src/actions/internal/check/index.js
@@ -1,3 +1,2 @@
-export * from './checkCommand.js';
 export * from './checkInstall.js';
 export * from './checkProject.js';